+++ UNITTESTS/moduletests/storage/kvstore/FileSystemStore/unittest.cmake
@@ -9,15 +9,15 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/HeapBlockDevice.cpp
-  ../features/storage/kvstore/filesystemstore/FileSystemStore.cpp
-  ../features/storage/filesystem/littlefs/LittleFileSystem.cpp
-  ../features/storage/filesystem/Dir.cpp
-  ../features/storage/filesystem/File.cpp
-  ../features/storage/filesystem/FileSystem.cpp
+  ../storage/blockdevice/HeapBlockDevice.cpp
+  ../storage/kvstore/filesystemstore/FileSystemStore.cpp
+  ../storage/filesystem/littlefs/LittleFileSystem.cpp
+  ../storage/filesystem/Dir.cpp
+  ../storage/filesystem/File.cpp
+  ../storage/filesystem/FileSystem.cpp
   ../features/frameworks/mbed-trace/source/mbed_trace.c
-  ../features/storage/filesystem/littlefs/littlefs/lfs_util.c
-  ../features/storage/filesystem/littlefs/littlefs/lfs.c
+  ../storage/filesystem/littlefs/littlefs/lfs_util.c
+  ../storage/filesystem/littlefs/littlefs/lfs.c
   ../platform/source/FileBase.cpp
   ../platform/source/FileSystemHandle.cpp
   ../platform/source/FileHandle.cpp
