+++ UNITTESTS/moduletests/storage/kvstore/TDBStore/moduletest.cpp
@@ -15,9 +15,9 @@
  */
 
 #include "gtest/gtest.h"
-#include "features/storage/blockdevice/HeapBlockDevice.h"
-#include "features/storage/blockdevice/FlashSimBlockDevice.h"
-#include "features/storage/kvstore/tdbstore/TDBStore.h"
+#include "storage/blockdevice/HeapBlockDevice.h"
+#include "storage/blockdevice/FlashSimBlockDevice.h"
+#include "storage/kvstore/tdbstore/TDBStore.h"
 #include <stdlib.h>
 
 #define BLOCK_SIZE (512)
