+++ UNITTESTS/storage/blockdevice/ExhaustibleBlockDevice/unittest.cmake
@@ -9,11 +9,11 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/FlashSimBlockDevice.cpp
+  ../storage/blockdevice/ExhaustibleBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/FlashSimBlockDevice/test_FlashSimBlockDevice.cpp
+  storage/blockdevice/ExhaustibleBlockDevice/test_ExhaustibleBlockDevice.cpp
 )
