+++ UNITTESTS/storage/blockdevice/FlashSimBlockDevice/test_FlashSimBlockDevice.cpp
@@ -15,12 +15,12 @@
  */
 
 #include "gtest/gtest.h"
-#include "features/storage/blockdevice/ExhaustibleBlockDevice.h"
+#include "storage/blockdevice/FlashSimBlockDevice.h"
 #include "stubs/BlockDevice_mock.h"
 
 #define BLOCK_SIZE (512)
 #define DEVICE_SIZE (BLOCK_SIZE*10)
-#define ERASE_CYCLES (2)
+#define ERASE_VALUE (2)
 
 using ::testing::_;
 using ::testing::Return;
@@ -32,12 +32,13 @@ using ::testing::SetArgPointee;
 using ::testing::SetArgReferee;
 using ::testing::DoAll;
 
-class ExhaustibleBlockModuleTest : public testing::Test {
+class FlashSimBlockModuleTest : public testing::Test {
 protected:
     BlockDeviceMock bd_mock;
-    ExhaustibleBlockDevice bd{&bd_mock, ERASE_CYCLES};
+    FlashSimBlockDevice bd{&bd_mock, ERASE_VALUE};
     uint8_t *magic;
     uint8_t *magic2;
+    uint8_t *erased_mem;
     uint8_t *buf;
     virtual void SetUp()
     {
@@ -51,11 +52,13 @@ protected:
         ASSERT_EQ(bd.init(), 0);
         magic = new uint8_t[BLOCK_SIZE];
         magic2 = new uint8_t[BLOCK_SIZE];
+        erased_mem = new uint8_t[BLOCK_SIZE];
         buf = new uint8_t[BLOCK_SIZE];
         // Generate simple pattern to verify against
         for (int i = 0; i < BLOCK_SIZE; i++) {
             magic[i] = 0xaa + i;
             magic2[i] = 0xaa + i + 1;
+            erased_mem[i] = ERASE_VALUE;
         }
     }
 
@@ -65,16 +68,16 @@ protected:
         ASSERT_EQ(bd.deinit(), 0);
         delete[] magic;
         delete[] magic2;
+        delete[] erased_mem;
         delete[] buf;
     }
 };
 
-TEST_F(ExhaustibleBlockModuleTest, init)
+TEST_F(FlashSimBlockModuleTest, init)
 {
-    ExhaustibleBlockDevice b(&bd_mock, ERASE_CYCLES);
+    FlashSimBlockDevice b(&bd_mock, ERASE_VALUE);
     EXPECT_EQ(b.get_erase_size(), 0);
     EXPECT_EQ(b.get_erase_size(0), 0);
-    EXPECT_EQ(b.get_erase_value(), BD_ERROR_DEVICE_ERROR);
     EXPECT_EQ(b.get_read_size(), 0);
     EXPECT_EQ(b.get_program_size(), 0);
     EXPECT_EQ(b.size(), 0);
@@ -92,44 +95,72 @@ TEST_F(ExhaustibleBlockModuleTest, init)
 
     EXPECT_EQ(b.get_erase_size(), bd_mock.get_erase_size());
     EXPECT_EQ(b.get_erase_size(0), bd_mock.get_erase_size(0));
-    EXPECT_EQ(b.get_erase_value(), bd_mock.get_erase_value());
+    EXPECT_EQ(b.get_erase_value(), ERASE_VALUE);
     EXPECT_EQ(b.get_program_size(), 512);
     EXPECT_EQ(b.get_read_size(), 512);
     EXPECT_EQ(b.size(), bd_mock.size());
     EXPECT_EQ(b.get_type(), "mytype");
+    EXPECT_EQ(b.sync(), 0);
     EXPECT_EQ(b.deinit(), 0);
+
 }
 
-TEST_F(ExhaustibleBlockModuleTest, program_unaligned)
+TEST_F(FlashSimBlockModuleTest, program_unaligned)
 {
     EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE-1), BD_ERROR_DEVICE_ERROR);
 }
 
-TEST_F(ExhaustibleBlockModuleTest, erase_cycle_limit_reached)
+TEST_F(FlashSimBlockModuleTest, program_no_erase)
 {
-    EXPECT_CALL(bd_mock, program(_, 0, BLOCK_SIZE))
+    // Data has not been erased, so this fails
+    EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), mbed::BD_ERROR_NOT_ERASED);
+
+    EXPECT_CALL(bd_mock, program(ByteBufferMatcher(erased_mem, BLOCK_SIZE), 0, BLOCK_SIZE))
     .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), 0);
+
+    EXPECT_CALL(bd_mock, program(ByteBufferMatcher(magic, BLOCK_SIZE), 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), 0);
+
+    EXPECT_CALL(bd_mock, read(_, 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(DoAll(SetArg0ToCharPtr(magic, BLOCK_SIZE), Return(BD_ERROR_OK)));
+
+    // No program() call should happen for this call.
+    EXPECT_EQ(bd.program(magic2, 0, BLOCK_SIZE), mbed::BD_ERROR_NOT_ERASED);
+}
+
+TEST_F(FlashSimBlockModuleTest, reprogram)
+{
+    EXPECT_CALL(bd_mock, program(ByteBufferMatcher(erased_mem, BLOCK_SIZE), 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), 0);
+
+    EXPECT_CALL(bd_mock, program(ByteBufferMatcher(magic, BLOCK_SIZE), 0, BLOCK_SIZE))
+    .Times(2)
     .WillRepeatedly(Return(BD_ERROR_OK));
 
     EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), 0);
 
-    EXPECT_CALL(bd_mock, erase(0, BLOCK_SIZE))
-    .Times(ERASE_CYCLES - 1) // Will fall silently after erase cycles are worn out.
-    .WillRepeatedly(Return(BD_ERROR_OK));
+    // Programming the same data.
+    EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), 0);
 
-    for (int i = 0; i < ERASE_CYCLES; i++) {
-        EXPECT_EQ(bd.erase(0, BLOCK_SIZE), 0);
-    }
+    EXPECT_CALL(bd_mock, program(ByteBufferMatcher(erased_mem, BLOCK_SIZE), 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
 
-    // Erase silently fails, no error report.
-    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), BD_ERROR_OK);
-    EXPECT_EQ(bd.program(magic2, 0, BLOCK_SIZE), BD_ERROR_OK);
+    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), 0);
 }
 
-TEST_F(ExhaustibleBlockModuleTest, erase_invalid)
+TEST_F(FlashSimBlockModuleTest, erase_invalid)
 {
-    ASSERT_GT(ERASE_CYCLES, 1);
-
     // Unaligned erase should fail
     EXPECT_EQ(bd.erase(0, BLOCK_SIZE-1), BD_ERROR_DEVICE_ERROR);
 }
