+++ UNITTESTS/storage/blockdevice/FlashSimBlockDevice/unittest.cmake
@@ -9,11 +9,11 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/MBRBlockDevice.cpp
+  ../storage/blockdevice/FlashSimBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/MBRBlockDevice/test_MBRBlockDevice.cpp
+  storage/blockdevice/FlashSimBlockDevice/test_FlashSimBlockDevice.cpp
 )
