+++ UNITTESTS/storage/blockdevice/HeapBlockDevice/unittest.cmake
@@ -9,13 +9,13 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/HeapBlockDevice.cpp
+  ../storage/blockdevice/HeapBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/HeapBlockDevice/test.cpp
+  storage/blockdevice/HeapBlockDevice/test.cpp
 )
 
 set(unittest-test-flags
