+++ UNITTESTS/storage/blockdevice/MBRBlockDevice/unittest.cmake
@@ -9,11 +9,11 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/ExhaustibleBlockDevice.cpp
+  ../storage/blockdevice/MBRBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/ExhaustibleBlockDevice/test_ExhaustibleBlockDevice.cpp
+  storage/blockdevice/MBRBlockDevice/test_MBRBlockDevice.cpp
 )
