+++ UNITTESTS/storage/blockdevice/ObservingBlockDevice/test_ObservingBlockDevice.cpp
@@ -15,36 +15,27 @@
  */
 
 #include "gtest/gtest.h"
+#include "storage/blockdevice/ObservingBlockDevice.h"
+#include "storage/blockdevice/ReadOnlyBlockDevice.h"
 #include "stubs/BlockDevice_mock.h"
-#include "features/storage/blockdevice/ReadOnlyBlockDevice.h"
-#include "platform/mbed_error.h"
+
+using ::testing::_;
+using ::testing::Return;
 
 #define BLOCK_SIZE (512)
 #define DEVICE_SIZE (BLOCK_SIZE*10)
 
-using ::testing::_;
-using ::testing::Return;
-using ::testing::DoAll;
-
-class ReadOnlyBlockModuleTest : public testing::Test {
+class ObservingBlockModuleTest : public testing::Test {
 protected:
     BlockDeviceMock bd_mock;
-    ReadOnlyBlockDevice bd{&bd_mock};
+    ObservingBlockDevice bd{&bd_mock};
     uint8_t *magic;
-    uint8_t *buf;
+    static int cnt;
     virtual void SetUp()
     {
-        ON_CALL(bd_mock, size()).WillByDefault(Return(DEVICE_SIZE));
-        ON_CALL(bd_mock, get_erase_size(_)).WillByDefault(Return(BLOCK_SIZE));
-        ON_CALL(bd_mock, get_erase_size()).WillByDefault(Return(BLOCK_SIZE));
-        ON_CALL(bd_mock, get_program_size()).WillByDefault(Return(BLOCK_SIZE));
-        ON_CALL(bd_mock, get_read_size()).WillByDefault(Return(BLOCK_SIZE));
-        ON_CALL(bd_mock, is_valid_erase(_, _)).WillByDefault(Return(true));
-        ON_CALL(bd_mock, init()).WillByDefault(Return(BD_ERROR_OK));
-
+        cnt = 0;
         ASSERT_EQ(bd.init(), 0);
         magic = new uint8_t[BLOCK_SIZE];
-        buf = new uint8_t[BLOCK_SIZE];
         // Generate simple pattern to verify against
         for (int i = 0; i < BLOCK_SIZE; i++) {
             magic[i] = 0xaa + i;
@@ -55,11 +46,16 @@ protected:
     {
         ASSERT_EQ(bd.deinit(), 0);
         delete[] magic;
-        delete[] buf;
+    }
+
+    static void bd_change(BlockDevice* bd) {
+        cnt++;
     }
 };
 
-TEST_F(ReadOnlyBlockModuleTest, init)
+int ObservingBlockModuleTest::cnt = 0;
+
+TEST_F(ObservingBlockModuleTest, init)
 {
     EXPECT_EQ(bd.get_erase_size(), bd_mock.get_erase_size());
     EXPECT_EQ(bd.get_erase_size(0), bd_mock.get_erase_size(0));
@@ -70,9 +66,28 @@ TEST_F(ReadOnlyBlockModuleTest, init)
     EXPECT_EQ(bd.get_type(), bd_mock.get_type());
 }
 
-TEST_F(ReadOnlyBlockModuleTest, write_protection) {
-    EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), MBED_ERROR_WRITE_PROTECTED);
-    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), MBED_ERROR_WRITE_PROTECTED);
+TEST_F(ObservingBlockModuleTest, program_erase_cb)
+{
+    EXPECT_CALL(bd_mock, program(magic, 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    bd.attach(this->bd_change);
+    EXPECT_EQ(bd.program(magic, 0, BLOCK_SIZE), 0);
+    EXPECT_EQ(cnt, 1);
+
+    EXPECT_CALL(bd_mock, erase(0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    EXPECT_EQ(bd.erase(0, BLOCK_SIZE), 0);
+    EXPECT_EQ(cnt, 2);
+
+    EXPECT_CALL(bd_mock, read(_, 0, BLOCK_SIZE))
+    .Times(1)
+    .WillOnce(Return(BD_ERROR_OK));
+
+    uint8_t buf[BLOCK_SIZE];
     EXPECT_EQ(bd.read(buf, 0, BLOCK_SIZE), 0);
-    EXPECT_EQ(bd.sync(), 0); // Should not have any influence
+    EXPECT_EQ(cnt, 2); // Counter stays the same
 }
