+++ UNITTESTS/storage/blockdevice/ProfilingBlockDevice/unittest.cmake
@@ -9,12 +9,13 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/ChainingBlockDevice.cpp
-  ../features/storage/blockdevice/HeapBlockDevice.cpp
+  ../storage/blockdevice/ProfilingBlockDevice.cpp
+  ../storage/blockdevice/HeapBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/ChainingBlockDevice/test_ChainingBlockDevice.cpp
+  storage/blockdevice/ProfilingBlockDevice/test_ProfilingBlockDevice.cpp
+  stubs/mbed_error.c
 )
