+++ UNITTESTS/storage/blockdevice/ReadOnlyBlockDevice/unittest.cmake
@@ -9,13 +9,13 @@ set(unittest-includes ${unittest-includes}
 )
 
 set(unittest-sources
-  ../features/storage/blockdevice/ReadOnlyBlockDevice.cpp
-  ../features/storage/blockdevice/HeapBlockDevice.cpp
+  ../storage/blockdevice/ReadOnlyBlockDevice.cpp
+  ../storage/blockdevice/HeapBlockDevice.cpp
   stubs/mbed_atomic_stub.c
   stubs/mbed_assert_stub.cpp
 )
 
 set(unittest-test-sources
-  features/storage/blockdevice/ReadOnlyBlockDevice/test_ReadOnlyBlockDevice.cpp
+  storage/blockdevice/ReadOnlyBlockDevice/test_ReadOnlyBlockDevice.cpp
   stubs/mbed_error.c
 )
