+++ UNITTESTS/stubs/BlockDevice_mock.h
@@ -20,7 +20,7 @@
 
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
-#include "features/storage/blockdevice/BlockDevice.h"
+#include "storage/blockdevice/BlockDevice.h"
 
 class BlockDeviceMock : public BlockDevice {
 public:
